abandon function approach
Dividing the processing of the data into functions became messy when passing different pieces of data back and forth, but mostly messy because handling the submit event is asynchronous.  Gathering the form data, creating the story html, then adding the story html to the DOM all need to be in one function to avoid using 'async', 'await', and promises.  This commit removes the extra functions creating only the one process(data) function.  The form data gathering, creating the story, and adding to the DOM will all be inside the submit event handler function within the outer process(data) function.